backend/requirements.txt
fastapi>=0.95
python-multipart>=0.0.6
uvicorn[standard]>=0.20
httpx>=0.24
openai>=1.0
sqlalchemy>=2.0
aiosqlite>=0.18
alembic>=1.11
pydantic-settings>=1.1
starlette>=0.27
bcrypt>=4.0
PyJWT>=2.8
websockets>=12.0
